[Auto-Fix] Fix ESLint & Tests failures from dependabot #46 (#47)
* chore(deps-dev): bump eslint

Bumps the development-dependencies group in /fs-crypto-coin-tracker/stage-2-node-express-mongo/back-end with 1 update: [eslint](https://github.com/eslint/eslint).


Updates `eslint` from 10.9.0 to 10.9.1
- [Release notes](https://github.com/eslint/eslint/releases)
- [Commits](https://github.com/eslint/eslint/compare/v10.9.0...v10.9.1)

---
updated-dependencies:
- dependency-name: eslint
  dependency-version: 10.9.1
  dependency-type: direct:development
  update-type: version-update:semver-patch
  dependency-group: development-dependencies
...


* fix: resolve ESLint violations from dependabot PR #46

* fix: support ESLint flat config


* fix: validate watchlist symbols


---------

+++ fs-crypto-coin-tracker/stage-2-node-express-mongo/back-end/modules/mongooseModule.js
@@ -1,33 +1,37 @@
 const WatchItem = require('../models/watchItem');
 
 const getWatchItemSymbols = async () => {
-    // fetch all watchlist docs from MongoDb
-    const watchItems = await WatchItem.find({})
+  // fetch all watchlist docs from MongoDb
+  const watchItems = await WatchItem.find({});
 
-    // we need to filter on symbol only..
-    const watchSymbols = watchItems.map(item => item.symbol);
+  // we need to filter on symbol only..
+  const watchSymbols = watchItems.map(item => item.symbol);
 
-    return watchSymbols;
-} 
+  return watchSymbols;
+};
 
 const deleteWatchItem = async (symbol) => {
-    // deleting the symbol from MongoDb
-    await WatchItem.deleteOne({ symbol: symbol });
-}
+  // deleting the symbol from MongoDb
+  if (typeof symbol !== 'string' || !/^[A-Za-z0-9]+$/.test(symbol)) {
+    throw new TypeError('Invalid symbol');
+  }
+
+  await WatchItem.deleteOne({ symbol: symbol });
+};
 const saveWatchItem = async (symbol) => {
     
-    // add symbol to watch list
-    const item = new WatchItem({
-        symbol: symbol,
-        dateCreated: Date.now()
-    })
+  // add symbol to watch list
+  const item = new WatchItem({
+    symbol: symbol,
+    dateCreated: Date.now()
+  });
 
-    // insert into mongoDb
-    await item.save();
-}
+  // insert into mongoDb
+  await item.save();
+};
 
 module.exports = {
-    getWatchItemSymbols,
-    deleteWatchItem,
-    saveWatchItem
-}
+  getWatchItemSymbols,
+  deleteWatchItem,
+  saveWatchItem
+};
